[librpbase] byteswap.h: Explicitly cast x to unsigned in the fallback __swab*() macros.
Reported by @DankRank.

+++ src/librpbase/byteswap.h
@@ -55,29 +55,29 @@
 #ifdef HAVE___BUILTIN_BSWAP16
 #define __swab16(x) ((uint16_t)__builtin_bswap16(x))
 #else
-#define __swab16(x) ((uint16_t)(((x) << 8) | ((x) >> 8)))
+#define __swab16(x) ((uint16_t)((((uint16_t)x) << 8) | (((uint16_t)x) >> 8)))
 #endif
 
 #ifdef HAVE___BUILTIN_BSWAP32
 #define __swab32(x) ((uint32_t)__builtin_bswap32(x))
 #else
 #define __swab32(x) \
-	((uint32_t)(((x) << 24) | ((x) >> 24) | \
-		(((x) & 0x0000FF00UL) << 8) | \
-		(((x) & 0x00FF0000UL) >> 8)))
+	((uint32_t)((((uint32_t)x) << 24) | (((uint32_t)x) >> 24) | \
+		((((uint32_t)x) & 0x0000FF00UL) << 8) | \
+		((((uint32_t)x) & 0x00FF0000UL) >> 8)))
 #endif
 
 #ifdef HAVE___BUILTIN_BSWAP64
 #define __swab64(x) ((uint64_t)__builtin_bswap64(x))
 #else
 #define __swab64(x) \
-	((uint64_t)(((x) << 56) | ((x) >> 56) | \
-		(((x) & 0x000000000000FF00ULL) << 40) | \
-		(((x) & 0x0000000000FF0000ULL) << 24) | \
-		(((x) & 0x00000000FF000000ULL) << 8) | \
-		(((x) & 0x000000FF00000000ULL) >> 8) | \
-		(((x) & 0x0000FF0000000000ULL) >> 24) | \
-		(((x) & 0x00FF000000000000ULL) >> 40)))
+	((uint64_t)((((uint64_t)x) << 56) | (((uint64_t)x) >> 56) | \
+		((((uint64_t)x) & 0x000000000000FF00ULL) << 40) | \
+		((((uint64_t)x) & 0x0000000000FF0000ULL) << 24) | \
+		((((uint64_t)x) & 0x00000000FF000000ULL) << 8) | \
+		((((uint64_t)x) & 0x000000FF00000000ULL) >> 8) | \
+		((((uint64_t)x) & 0x0000FF0000000000ULL) >> 24) | \
+		((((uint64_t)x) & 0x00FF000000000000ULL) >> 40)))
 #endif
 
 #endif /* defined(_MSC_VER) */
